+++ package.json
@@ -3,6 +3,7 @@
   "version": "27.0.0",
   "private": true,
   "description": "Личный проект «Кексобукинг» от HTML Academy",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git@github.com:htmlacademy-javascript/2118329-keksobooking-27.git"
